Rename protocols -> broadcast. Add #*# to .gitignore
Renamed use_fft -> use_omega_powers

remove DS_Store

Move task_pool into utils

renaming, compressing for reed solomon

rename pid -> sid in program runner

Renaming of offline / randousha

utils refactored

renames apps/shuffle to apps/aynchromix
creates combined tests/test_asynchromix.py
