+++ apps/asynchromix/cpp/compute-power-sums.cpp
