+++ apps/asynchromix/powermixing.py
@@ -98,7 +98,7 @@ async def phase3(context, **kwargs):
 
 async def async_mixing(n, t, k):
     from .solver.solver import solve
-    from honeybadgermpc.task_pool import TaskPool
+    from honeybadgermpc.utils.task_pool import TaskPool
 
     pr1 = TaskProgramRunner(n, t)
     file_prefixes = [uuid.uuid4().hex for _ in range(k)]
@@ -133,7 +133,7 @@ async def build_powermixing_cpp_code():
 async def async_mixing_in_processes(network_info, n, t, k, run_id, node_id):
     from .solver.solver import solve
     from honeybadgermpc.ipc import ProcessProgramRunner
-    from honeybadgermpc.task_pool import TaskPool
+    from honeybadgermpc.utils.task_pool import TaskPool
 
     file_prefixes = [uuid.uuid4().hex for _ in range(k)]
     async with ProcessProgramRunner(network_info, n, t, node_id) as runner:
