+++ apps/asynchromix/solver/solver.cpp
