+++ honeybadgermpc/broadcast/commonsubset.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+import time
 
 
 async def commonsubset(pid, n, f, rbc_out, aba_in, aba_out):
@@ -67,9 +68,9 @@ async def commonsubset(pid, n, f, rbc_out, aba_in, aba_out):
 
 
 async def make_commonsubset(sid, pid, n, f, pk, sk, input_msg, send, recv, bcast):
-    from honeybadgermpc.protocols.commoncoin import shared_coin
-    from honeybadgermpc.protocols.binaryagreement import binaryagreement
-    from honeybadgermpc.protocols.reliablebroadcast import reliablebroadcast
+    from honeybadgermpc.broadcast.commoncoin import shared_coin
+    from honeybadgermpc.broadcast.binaryagreement import binaryagreement
+    from honeybadgermpc.broadcast.reliablebroadcast import reliablebroadcast
 
     coin_recvs = [asyncio.Queue() for _ in range(n)]
     aba_recvs = [asyncio.Queue() for _ in range(n)]
@@ -166,6 +167,7 @@ async def run_common_subset_in_processes(config, pbk, pvk, n, f, nodeid):
             make_commonsubset(
                 sid, nodeid, n, f, pbk, pvk, input_q.get, send, recv, bcast))
 
+        start_time = time.time()
         await input_q.put('<[ACS Input %d]>' % nodeid)
         acs, recv_tasks, work_tasks = await create_acs_task
         acs_output = await acs
@@ -174,6 +176,7 @@ async def run_common_subset_in_processes(config, pbk, pvk, n, f, nodeid):
             task.cancel()
 
         logging.info(f"OUTPUT: {acs_output}")
+        logging.info(f"Elapsed time: {time.time() - start_time}")
 
 
 if __name__ == "__main__":
@@ -181,8 +184,8 @@ if __name__ == "__main__":
     import base64
     from honeybadgermpc.config import HbmpcConfig
     from honeybadgermpc.ipc import ProcessProgramRunner
-    from honeybadgermpc.protocols.crypto.boldyreva import TBLSPublicKey  # noqa:F401
-    from honeybadgermpc.protocols.crypto.boldyreva import TBLSPrivateKey  # noqa:F401
+    from honeybadgermpc.broadcast.crypto.boldyreva import TBLSPublicKey  # noqa:F401
+    from honeybadgermpc.broadcast.crypto.boldyreva import TBLSPrivateKey  # noqa:F401
 
     pbk = pickle.loads(base64.b64decode(HbmpcConfig.extras["public_key"]))
     pvk = pickle.loads(base64.b64decode(HbmpcConfig.extras["private_key"]))
