+++ honeybadgermpc/progs/mixins/dataflow.py
@@ -2,7 +2,7 @@ import asyncio
 from abc import ABC, abstractmethod
 from honeybadgermpc.field import GFElement
 from honeybadgermpc.progs.mixins.constants import MixinConstants
-from honeybadgermpc.utils import type_check
+from honeybadgermpc.utils.typecheck import type_check
 
 
 class GFElementFuture(ABC, asyncio.Future):
