+++ honeybadgermpc/robust_reconstruction.py
@@ -7,9 +7,11 @@ from honeybadgermpc.batch_reconstruction import fetch_one
 
 
 async def robust_reconstruct(field_futures, field, n, t, point):
-    use_fft = point.use_fft
-    enc = EncoderFactory.get(point, Algorithm.FFT if use_fft else Algorithm.VANDERMONDE)
-    dec = DecoderFactory.get(point, Algorithm.FFT if use_fft else Algorithm.VANDERMONDE)
+    use_omega_powers = point.use_omega_powers
+    enc = EncoderFactory.get(point, Algorithm.FFT if use_omega_powers
+                             else Algorithm.VANDERMONDE)
+    dec = DecoderFactory.get(point, Algorithm.FFT if use_omega_powers
+                             else Algorithm.VANDERMONDE)
     robust_dec = RobustDecoderFactory.get(t, point, algorithm=Algorithm.GAO)
     incremental_decoder = IncrementalDecoder(enc, dec, robust_dec, t, 1, t)
