+++ tests/crypto/test_boldyreva.py
@@ -2,13 +2,13 @@ import pickle
 import random
 from base64 import encodebytes
 from pytest import mark
-from honeybadgermpc.protocols.crypto.boldyreva import dealer
+from honeybadgermpc.broadcast.crypto.boldyreva import dealer
 
 
 class TestTBLSPublicKey:
 
     def test_init(self, vk, vks):
-        from honeybadgermpc.protocols.crypto.boldyreva import TBLSPublicKey
+        from honeybadgermpc.broadcast.crypto.boldyreva import TBLSPublicKey
         players = 10    # TODO bind to fixtures
         count = 5   # TODO bind to fixtures
         public_key = TBLSPublicKey(players, count, vk, vks)
@@ -28,7 +28,7 @@ class TestTBLSPublicKey:
         assert tbls_public_key.__dict__ == original_dict
 
     def test_setstate(self, tbls_public_key, serialized_tbls_public_key_dict):
-        from honeybadgermpc.protocols.crypto.boldyreva import TBLSPublicKey
+        from honeybadgermpc.broadcast.crypto.boldyreva import TBLSPublicKey
         unset_public_key = TBLSPublicKey(None, None, None, None)
         unset_public_key.__setstate__(serialized_tbls_public_key_dict)
         assert len(unset_public_key.__dict__) == len(tbls_public_key.__dict__)
@@ -65,7 +65,7 @@ def test_boldyreva():
 
 @mark.parametrize('n', (0, 1, 2))
 def test_deserialize_arg(n, g, mocker):
-    from honeybadgermpc.protocols.crypto import boldyreva
+    from honeybadgermpc.broadcast.crypto import boldyreva
     mocked_deserialize = mocker.patch.object(
         boldyreva.group, 'deserialize', autospec=True)
     deserialize_func = getattr(boldyreva, 'deserialize{}'.format(n))
