+++ tests/test_avss_value_processor.py
@@ -3,7 +3,7 @@ from pytest import mark
 from contextlib import ExitStack
 from pickle import dumps
 from honeybadgermpc.avss_value_processor import AvssValueProcessor
-from honeybadgermpc.protocols.crypto.boldyreva import dealer
+from honeybadgermpc.broadcast.crypto.boldyreva import dealer
 
 
 @mark.asyncio
