+++ tests/test_reed_solomon_wb.py
@@ -1,6 +1,6 @@
 import random
 import pytest
-from honeybadgermpc.wb_interpolate import make_wb_encoder_decoder
+from honeybadgermpc.reed_solomon_wb import make_wb_encoder_decoder
 
 
 def test_decoding():
