+++ tests/test_utils.py
@@ -1,5 +1,8 @@
-from pytest import raises
-from honeybadgermpc.utils import static_type_check, wrap_send
+from pytest import raises, mark
+from honeybadgermpc.utils.typecheck import static_type_check
+from honeybadgermpc.utils.misc import wrap_send
+from random import randint
+import asyncio
 
 
 def test_type_check_callable():
@@ -139,3 +142,22 @@ def test_wrap_send():
     wrapped = wrap_send('hello', _send)
     wrapped(1, 'world')
     assert (test_dest, test_message) == (1, ('hello', 'world'))
+
+
+@mark.asyncio
+async def test_pool():
+    from honeybadgermpc.utils.task_pool import TaskPool
+
+    max_tasks = 4
+
+    async def work(q, max_tasks):
+        assert q.qsize() <= max_tasks
+        q.put_nowait(None)
+        await asyncio.sleep(randint(0, 3)*0.2)
+        await q.get()
+
+    pool = TaskPool(max_tasks)
+    q = asyncio.Queue()
+    for _ in range(10):
+        pool.submit(work(q, max_tasks))
+    await pool.close()
